Create README.md
Time Series analysis is a method to understand the behaviour of a particular variable with time and hence predict the future model of that variable (with the help of its data) with respect to time.
In this project I tried to understand the behaviour of Sulphur Dioxide content in the atmosphere of the Delhi with the help of a dataset (source: Kaggle) which provided the information of Sulphur Dioxide content in the air of Delhi from 1987 to 2015. And hence predict the air quality of Delhi  till the year 2019 with the help of ARIMA model. The project was done on R.